+++ src/Symfony/Component/Dotenv/Tests/DotenvTest.php
@@ -208,10 +208,79 @@ class DotenvTest extends TestCase
     public function testEnvVarIsNotOverriden()
     {
         putenv('TEST_ENV_VAR=original_value');
+        $_SERVER['TEST_ENV_VAR'] = 'original_value';
 
         $dotenv = new DotEnv();
         $dotenv->populate(array('TEST_ENV_VAR' => 'new_value'));
 
         $this->assertSame('original_value', getenv('TEST_ENV_VAR'));
+    }
+
+    public function testHttpVarIsPartiallyOverriden()
+    {
+        $_SERVER['HTTP_TEST_ENV_VAR'] = 'http_value';
+
+        $dotenv = new DotEnv();
+        $dotenv->populate(array('HTTP_TEST_ENV_VAR' => 'env_value'));
+
+        $this->assertSame('env_value', getenv('HTTP_TEST_ENV_VAR'));
+        $this->assertSame('env_value', $_ENV['HTTP_TEST_ENV_VAR']);
+        $this->assertSame('http_value', $_SERVER['HTTP_TEST_ENV_VAR']);
+    }
+
+    public function testMemorizingLoadedVarsNamesInSpecialVar()
+    {
+        // Special variable not exists
+        unset($_ENV['SYMFONY_DOTENV_VARS']);
+        unset($_SERVER['SYMFONY_DOTENV_VARS']);
+        putenv('SYMFONY_DOTENV_VARS');
+
+        unset($_ENV['APP_DEBUG']);
+        unset($_SERVER['APP_DEBUG']);
+        putenv('APP_DEBUG');
+        unset($_ENV['DATABASE_URL']);
+        unset($_SERVER['DATABASE_URL']);
+        putenv('DATABASE_URL');
+
+        $dotenv = new DotEnv();
+        $dotenv->populate(array('APP_DEBUG' => '1', 'DATABASE_URL' => 'mysql://root@localhost/db'));
+
+        $this->assertSame('APP_DEBUG,DATABASE_URL', getenv('SYMFONY_DOTENV_VARS'));
+
+        // Special variable has a value
+        $_ENV['SYMFONY_DOTENV_VARS'] = 'APP_ENV';
+        $_SERVER['SYMFONY_DOTENV_VARS'] = 'APP_ENV';
+        putenv('SYMFONY_DOTENV_VARS=APP_ENV');
+
+        $_ENV['APP_DEBUG'] = '1';
+        $_SERVER['APP_DEBUG'] = '1';
+        putenv('APP_DEBUG=1');
+        unset($_ENV['DATABASE_URL']);
+        unset($_SERVER['DATABASE_URL']);
+        putenv('DATABASE_URL');
+
+        $dotenv = new DotEnv();
+        $dotenv->populate(array('APP_DEBUG' => '0', 'DATABASE_URL' => 'mysql://root@localhost/db'));
+        $dotenv->populate(array('DATABASE_URL' => 'sqlite:///somedb.sqlite'));
+
+        $this->assertSame('APP_ENV,DATABASE_URL', getenv('SYMFONY_DOTENV_VARS'));
+    }
+
+    public function testOverridingEnvVarsWithNamesMemorizedInSpecialVar()
+    {
+        putenv('SYMFONY_DOTENV_VARS=FOO,BAR,BAZ');
+
+        putenv('FOO=foo');
+        putenv('BAR=bar');
+        putenv('BAZ=baz');
+        putenv('DOCUMENT_ROOT=/var/www');
+
+        $dotenv = new DotEnv();
+        $dotenv->populate(array('FOO' => 'foo1', 'BAR' => 'bar1', 'BAZ' => 'baz1', 'DOCUMENT_ROOT' => '/boot'));
+
+        $this->assertSame('foo1', getenv('FOO'));
+        $this->assertSame('bar1', getenv('BAR'));
+        $this->assertSame('baz1', getenv('BAZ'));
+        $this->assertSame('/var/www', getenv('DOCUMENT_ROOT'));
     }
 }